requirements.txt
Django==1.8.3
MySQL-python==1.2.5
argparse==1.2.1
djangorestframework==3.1.3
ipdb==0.8.1
ipython==3.2.1
wsgiref==0.1.2
